ADDED: TextureComponent; ADDED: dynamically applying of textures working - on cube; MODIFIED: minor refactoring;

+++ public_html/js/engine/ShaderChunk/fragment_color_para.glsl
@@ -1,1 +1,1 @@
-varying lowp vec4 vVertexColor;
+varying highp vec4 vVertexColor;
