+++ public_html/js/engine/ShaderChunk/fragment_texture_sampler.glsl
@@ -1,1 +1,1 @@
-mediump vec4 texColor = texture2D(uSampler, vec2(vTextureCoord.s, vTextureCoord.t));
+highp vec4 texColor = texture2D(uSampler, vec2(vTextureCoord.s, vTextureCoord.t));
