+++ public_html/js/engine/ShaderChunk/vertex_normal_para.glsl
@@ -1,3 +1,2 @@
-attribute lowp vec3 aVertexNormal;
-
-uniform lowp mat4 uNormalMatrix;
+attribute highp vec3 aVertexNormal;
+uniform highp mat4 uNormalMatrix;
